refactor(bottom-scroll-sheet): separate feature folder
