+++ lib/features/bottom_scroll_sheet/bottom_scroll_sheet.dart
@@ -3,11 +3,11 @@ import "dart:math";
 import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 
-import "../../../../config/map_view_config.dart";
-import "../../../../theme/app_theme.dart";
-import "../../controllers/bottom_sheet_controller.dart";
-import "../../controllers/controllers_set.dart";
-import "../map_config.dart";
+import "../../config/map_view_config.dart";
+import "../../theme/app_theme.dart";
+import "../map_view/controllers/bottom_sheet_controller.dart";
+import "../map_view/controllers/controllers_set.dart";
+import "../map_view/widgets/map_config.dart";
 import "map_view_pop_behaviour.dart";
 import "sheet_layout_scheme.dart";
