+++ lib/features/bottom_scroll_sheet/navigate_button.dart
@@ -1,12 +1,12 @@
 import "package:flutter/material.dart";
 import "package:flutter_riverpod/flutter_riverpod.dart";
 
-import "../../../../config/map_view_config.dart";
-import "../../../../theme/app_theme.dart";
-import "../../../../utils/context_extensions.dart";
-import "../../../parkings_view/widgets/parkings_icons.icons.dart";
-import "../../controllers/controllers_set.dart";
-import "../map_config.dart";
+import "../../config/map_view_config.dart";
+import "../../theme/app_theme.dart";
+import "../../utils/context_extensions.dart";
+import "../map_view/controllers/controllers_set.dart";
+import "../map_view/widgets/map_config.dart";
+import "../parkings_view/widgets/parkings_icons.icons.dart";
 
 class NavigateButton<T extends GoogleNavigable> extends ConsumerWidget {
   const NavigateButton({super.key});
