+++ lib/features/map_view/map_view.dart
@@ -6,10 +6,10 @@ import "package:flutter_riverpod/flutter_riverpod.dart";
 
 import "../../config/map_view_config.dart";
 import "../../theme/app_theme.dart";
+import "../bottom_scroll_sheet/bottom_scroll_sheet.dart";
+import "../bottom_scroll_sheet/sheet_layout_scheme.dart";
 import "controllers/bottom_sheet_controller.dart";
 import "controllers/controllers_set.dart";
-import "widgets/bottom_scroll_sheet/bottom_scroll_sheet.dart";
-import "widgets/bottom_scroll_sheet/sheet_layout_scheme.dart";
 import "widgets/map_config.dart";
 import "widgets/map_widget.dart";
